laravel training 会員登録、ログイン、メール、認証周りの処理追加

+++ src/Laravel9TaskList/resources/views/tasks/index.blade.php
@@ -13,6 +13,7 @@
                             </a>
                         </div>
                     </div>
+                    <div class="panel-body h4 pb-2 mb-4">あなたのタスク</div>
                     <table class="table">
                         <thead>
                             <tr>
@@ -26,16 +27,45 @@
                         </thead>
                         <tbody>
                             @foreach($tasks as $task)
-                                <tr>
-                                    <td class="title">{{ $task->title }}</td>
-                                    <td class="description">{{ $task->description }}</td>
-                                    <td>
-                                        <span class="label {{ $task->status_class }}">{{ $task->status_label }}</span>
-                                    </td>
-                                    <td>{{ $task->format_due_date }}</td>
-                                    <td><a href="/tasks/{{ $task->id }}/edit">編集</a></td>
-                                    <td><a href="/tasks/{{ $task->id }}/delete">削除</a></td>
-                                </tr>
+                                @if($task->user_id === Auth::user()->id)
+                                    <tr>
+                                        <td class="title">{{ $task->title }}</td>
+                                        <td class="description">{{ $task->description }}</td>
+                                        <td>
+                                            <span class="label {{ $task->status_class }}">{{ $task->status_label }}</span>
+                                        </td>
+                                        <td>{{ $task->format_due_date }}</td>
+                                        <td><a href="/tasks/{{ $task->id }}/edit">編集</a></td>
+                                        <td><a href="/tasks/{{ $task->id }}/delete">削除</a></td>
+                                    </tr>
+                                @endif
+                            @endforeach
+                        </tbody>
+                    </table>
+                    <div class="panel-body h4 pb-2 mb-4">他人のタスク</div>
+                    <table class="table">
+                        <thead>
+                            <tr>
+                                <th>タイトル</th>
+                                <th>説明文</th>
+                                <th>状態</th>
+                                <th>期限</th>
+                                <th>ユーザー名</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            @foreach($tasks as $task)
+                                @if($task->user_id !== Auth::user()->id)
+                                    <tr>
+                                        <td class="title">{{ $task->title }}</td>
+                                        <td class="description">{{ $task->description }}</td>
+                                        <td>
+                                            <span class="label {{ $task->status_class }}">{{ $task->status_label }}</span>
+                                        </td>
+                                        <td>{{ $task->format_due_date }}</td>
+                                        <td>{{ $task->user_name }}</td>
+                                    </tr>
+                                @endif
                             @endforeach
                         </tbody>
                     </table>
